+++ src/App.css
@@ -38,94 +38,3 @@
 } */
 
 
-.App{
-  background-image: url('../src/images/bimg.jpg');
-  background-size: cover;
-  background-repeat: no-repeat;
-  height:100vh;
-  width:100vw;
-  align-items: center;
-  justify-content: center;
-  display: flex; 
-  
-
-}
-.contain{
-  width:700px;
-  height :250px;
- /* margin-left: auto;
-  margin-right: auto;*/
-  background-color: white;
-  opacity:0.8;
-  align-items: center;
-  margin: auto;
-  padding-top: 250px;
-  
-  
-}
-.image{
-  width: 50px;
-  height: 50px;
-  float: left;
-  margin-top: -200px;
-  
-}
-.dh2{
-  float:left;
-  margin-left:60px;
-  margin-top:-215px;
-}
-.table{
-  float: left; 
-  /*align-items: center; */
-  /*justify-content: center;
-  display: flex;*/ 
-  padding-left: 40px;
-  background-color: beige;
-  border-radius:10px;
-  margin-left:11.5%;
-  margin-top:-100px;
-  width:500px;
-  height:250px;
-}
-.body{
-  /*background-image: url('../../../../Fcode Labs/Diary Card/sign-in/src/image/bimg.jpg');
-  background-size: cover;
-  background-repeat: no-repeat;*/
-}
-
-#txt1{
-  width:300px;
-  height:50px;
-  border:3px solid white;
-  border-radius:10px;
-}
-#btn1{
-  width:125px;
-  height:50px;
-  /*padding-left:20px;*/
-  margin-left:15px;
-  margin-top:10px;
-  border:2px solid rgb(1, 1, 49);
-  color:white;
-  background-color: #f34e1c;
-  border-radius:10px;
-}
-#btn1:hover{
-  background-color: #681a03;
-}
-#btn2{
-  width:125px;
-  height:50px;
-  /*padding-left:20px;*/
-  margin-left:170px;
-  /*align-items:center;*/
-  margin-top:20px;
-  border:2px solid rgb(1, 1, 49);
-  color:white;
-  background-color: #f34e1c;
-  border-radius:10px;
-}
-#btn2:hover{
-  background-color: #681a03;
-}
