--- src/signin.js
@@ -1,6 +1,0 @@
-export function genName() 
-{
-    console.log("asdf")
-    
-
-}
